moved the search stuff into a separate folder and got rid of the magnifying glass file
